Revert "Allow cherry-picks to handle merge commits (#7396)"
This reverts commit 2c685f353c21ffc0b7e62c301547b19f97a513aa.
A revert is necessary because it's not working without further changes.
